+++ src/constant.h
@@ -388,6 +388,7 @@ constexpr size_t SYS_LIMIT_FUNC_LEN = UINT32_MAX;
 constexpr size_t SYS_LIMIT_NATIVE_RECURSION = 256;
 constexpr size_t SYS_LIMIT_FUNC_DEPTH = 32;
 constexpr size_t SYS_LIMIT_UPVAR_NUM = UINT8_MAX;
+constexpr size_t SYS_LIMIT_JOB_DESC_LEN = 96;
 
 } // namespace ydsh