+++ src/object.cpp
@@ -18,6 +18,7 @@
 
 #include "core.h"
 #include "misc/files.h"
+#include "misc/format.hpp"
 #include "misc/num_util.hpp"
 #include "redir.h"
 #include "vm.h"
@@ -965,6 +966,21 @@ TimerObject::~TimerObject() {
   }
   fputs(out.c_str(), stderr);
   fflush(stderr);
+}
+
+bool formatJobDesc(const StringRef ref, std::string &out) {
+  return splitByDelim(ref, '\n', [&out](StringRef sub, bool newline) {
+    out += sub;
+    if (newline) {
+      out += "\\n";
+    }
+    if (out.size() > SYS_LIMIT_JOB_DESC_LEN) {
+      out.resize(SYS_LIMIT_JOB_DESC_LEN - 3);
+      out += "...";
+      return false;
+    }
+    return true;
+  });
 }
 
 } // namespace ydsh