+++ src/object.h
@@ -1323,6 +1323,16 @@ public:
   ~TimerObject();
 };
 
+/**
+ * for JobObject
+ * @param ref
+ * @param out
+ * append formatted string
+ * @return
+ * if out.size() reaches limit, trim out and return false
+ */
+bool formatJobDesc(StringRef ref, std::string &out);
+
 }; // namespace ydsh
 
 #endif // YDSH_OBJECT_H