+++ CHANGELOG.md
@@ -24,6 +24,8 @@
     - ``SIGSTOP``, ``SIGTERM``, ``SIGTSTP``, ``SIGTTIN``, ``SIGTTOU``, ``SIGUSR1``, ``SIGUSR2``
 - add ``String#sanitize`` method
     - replace invalid utf8 bytes and null characters
+- add builtin ``jobs`` command
+    - show job information (except for disowned jobs)
 
 #### API
 
@@ -59,6 +61,7 @@
     - always disable job-control in command substitution
     - propagate IO error as ``SystemError``
     - cancel command substitution by SIGINT
+- show job information via CTRL-Z/``fg``/``bg``
 - escape unprintable character when throw some errors
 
 #### Builtin